+++ src/contracts/ImplicitAccount/index.ts
@@ -1,0 +1,2 @@
+import ImplicitAccount from './container';
+export default ImplicitAccount;
